return to instance methods; rework to make structure clear and concise
