--- api/processors/group/actions/__init__.py
@@ -1,7 +1,0 @@
-from .abort import AbortProcessor
-from .start import StartProcessor
-
-__all__ = [
-    "AbortProcessor",
-    "StartProcessor",
-]
